feat(RHTAPWATCH-764): task for allocating ephemeral namespaces
We need to be able to run integration tests for components
built in Konflux. for this, we need an ephemeral environment
that will be completely clean of previous possible artifacts,
in which we can run the tests in, and remove afterwards.

Here we create a Tekton task that allocates an ephemeral
namespace for a single test, using the sandbox components.

Using a Custom resource, sandbox API creates a namespace
and supply the login information for it with Tekton results.
once the Pipelinerun is removed - the namespace is removed along
with it as the PipelineRun is configured as the owner of the
task with an ownerReference.

testing demo for allocating SpaceRequest

Signed-off-by: Omer <oamsalem@redhat.com>