chore: update build-definitions pipeline readmes
Now that we have a task that is included in a pipeline template, we
need to regenerate the pipelines readme in case we change the task
parameters' description.

For that we need to run the relevant script, and we need the script
to run with an appropriate image.

This is the second time trying this change as the first one failed
to run due to permissions issues. This time, we use the same image
for the git commands and to run the script and hopefully that should
address the previous issue.

Signed-off-by: Yftach Herzog <yherzog@redhat.com>