chore(KFLUXVNGD-555): migrate to 0.2.1
We already merged a migration script and bumped
the task version in previous PR.
However, tkn-bundle had a bug that
prevented us from adding the needed annotation.

A fix should have been done to tkn-bundle task
to succeed in the migration.
Now that the fix is in place we are updating
the task file and the migration file
to build the new version of the task.

Signed-off-by: Avi Biton <abiton@redhat.com>